Thêm CRUD: cho ChuyenKhoan

+++ QuanLyNganHang/ET/ET_KhoanVay.cs
@@ -6,7 +6,7 @@ namespace ET
 {
     public class ET_KhoanVay
     {
-        public ET_KhoanVay(string mAVAY, float sOTIENVAY, DateTime nGAYVAY, DateTime tHOIHAN, string tRANGTHAI, string mAKH, string mALAISUAT)
+        public ET_KhoanVay(string mAVAY, decimal sOTIENVAY, DateTime nGAYVAY, DateTime tHOIHAN, string tRANGTHAI, string mAKH, string mALAISUAT)
         {
             MAVAY = mAVAY;
             SOTIENVAY = sOTIENVAY;
@@ -18,7 +18,7 @@ namespace ET
         }
 
         public string MAVAY { get; set; }
-        public float SOTIENVAY { get; set; }
+        public decimal SOTIENVAY { get; set; }
         public DateTime NGAYVAY { get; set; }
         public DateTime THOIHAN { get; set; }
         public string TRANGTHAI { get; set; }
